Move database tokens selection to selection module.
