+++ modules/selection/src/main/kotlin/com.r3.corda.lib.tokens.selection/memory/config/InMemorySelectionConfig.kt
@@ -1,22 +1,16 @@
-package com.r3.corda.lib.tokens.selection.config
+package com.r3.corda.lib.tokens.selection.memory.config
 
 import co.paralleluniverse.fibers.Suspendable
-import com.r3.corda.lib.tokens.selection.selectors.LocalTokenSelector
-import com.r3.corda.lib.tokens.selection.selectors.Selector
-import com.r3.corda.lib.tokens.selection.services.VaultWatcherService
+import com.r3.corda.lib.tokens.selection.api.StateSelectionConfig
+import com.r3.corda.lib.tokens.selection.memory.selector.LocalTokenSelector
+import com.r3.corda.lib.tokens.selection.memory.services.VaultWatcherService
 import net.corda.core.cordapp.CordappConfig
 import net.corda.core.cordapp.CordappConfigException
 import net.corda.core.node.ServiceHub
 import org.slf4j.LoggerFactory
 
 const val CACHE_SIZE_DEFAULT = 1024 // TODO Return good default, for now it's not wired, it will be done in separate PR.
-val INDEXING_STRATEGY_DEFAULT = VaultWatcherService.IndexingType.TOKEN // Default is by token class and identifier
-
-// TODO that interface should be separate, same for selector
-interface StateSelectionConfig {
-    @Suspendable
-    fun toSelector(services: ServiceHub): Selector
-}
+val INDEXING_STRATEGY_DEFAULT = VaultWatcherService.IndexingType.TOKEN_ONLY // Default is by token class and identifier
 
 data class InMemorySelectionConfig(val indexingStrategy: VaultWatcherService.IndexingType, val cacheSize: Int = CACHE_SIZE_DEFAULT) : StateSelectionConfig {
     companion object {
